Remove git-utils
