+++ packages/scafflater/util/run-command.js
